requirements.txt
Flask
flask-cors
pandas
python-Levenshtein
openpyxl